+++ client/src/view/Admin/Academic_Articles/academic_article_edit.vue
@@ -1,18 +1,18 @@
 <template>
   <div class="container">
     <center>
-      <h4 class="header">แก้ไขบทความวิชาการ</h4>
+      <h4 class="header">แก้ไข บทความวิชาการ</h4>
     </center>
     <div class="row mt-5" v-if="thisAcademic_Article && the_user">
       <div class="col-lg-2 col-xs-12"></div>
       <div class="col-lg-8 col-xs-12">
-        <img class="admin-img" v-if="url" :src="url" />
-        <img class="admin-img" v-else :src="getImgUrl(thisAcademic_Article.aa_image)" />
+        <img class=" " v-if="url" :src="url"  width="100%"/>
+        <img class=" " v-else :src="getImgUrl(thisAcademic_Article.aa_image)" width="100%"/>
         <br />
         <br />
         <button
           type="button"
-          class="form-control btn-success col-lg-12"
+          class="form-control btn-success btn col-lg-6 block-center"
           @click="ChooseFilesImage"
         >เลือกรูป</button>
         <br />
@@ -51,13 +51,13 @@
             rows="6"
           ></textarea>
           <br />
-          <div v-if="thisFiles != null" v-for="(file,run) in thisFiles">
+          <div v-if="academic_article_data_file != null" v-for="(file,run) in academic_article_data_file">
             <a :href="loadFile(file.f_name)" download>Dowload File</a>
             {{file.f_title}}
             <button
               type="button"
               class="btn btn-danger col-lg-1"
-              @click="RemoveFile(file.f_id)"
+              @click="RemoveFile(file.f_id,file.f_title)"
             >ลบ</button>
             <br />
             <br />
@@ -116,6 +116,7 @@
   </div>
 </template>
 <script>
+import axios from "axios";
 export default {
   data() {
     return {
@@ -124,7 +125,16 @@ export default {
       url: null,
       fileimage: "",
       files: [],
-      max_size_file: 0
+      max_size_file: 0,
+
+      academic_article_category_all:'',
+      data_load_category:false,
+      
+      academic_article_data:'',
+      academic_article_data_file:'',
+      data_load:false,
+      check_delete_file:false
+
     };
   },
   methods: {
@@ -176,11 +186,32 @@ export default {
       this.file_title.splice(index, 1);
     },
     // delete file
-    RemoveFile: function(fileID) {
+    RemoveFile: function(fileID,fileName) {
       var FD_delete = new FormData();
       FD_delete.append("fileID", fileID);
       FD_delete.append("creator", JSON.stringify(this.$store.state.log_on));
-      this.$store.dispatch("Delete_File", FD_delete);
+
+      this.$swal({
+          title: "Are you sure?",
+          text: "You want delete this Filse Name "+fileName,
+          icon: "warning",
+          buttons: true,
+          dangerMode: true,
+        })
+        .then((willDelete) => {
+        if (willDelete) {
+          this.$store.dispatch("Delete_File", FD_delete);
+          setTimeout(() => {
+            this.check_delete_file=true
+            this.data_load=false
+          },100);
+          swal({title: "Delete Success.",icon: "success",});
+        } else {
+          //
+        }
+      })
+      
+      
     },
     // submit
     submitArticle() {
@@ -210,34 +241,33 @@ export default {
       return this.$store.getters.getPath_Files;
     },
     thisAcademic_Article() {
-      var academic_articleAll = this.$store.getters.getAcademic_Article;
-      var academic_article;
-      for (var i = 0; i < academic_articleAll.length; i++) {
-        if (
-          academic_articleAll[i].aa_id == this.$route.params.AcademicArticleID
-        ) {
-          academic_article = academic_articleAll[i];
-        }
+      var academic_articleID = this.$route.params.AcademicArticleID;
+      if(this.data_load==false){
+        axios.get(this.$store.getters.getBase_Url+'Academic_article/get_this_academic_article/'+academic_articleID)
+        .then(response => {
+            // console.log(response.data)
+            this.academic_article_data = response.data[0][0],
+            this.academic_article_data_file = response.data[1]
+        })
+        this.data_load = true
       }
-      this.academic_articleE = academic_article;
-      return academic_article;
-    },
-    thisFiles() {
-      var filesAll = this.$store.getters.getFiles;
-      var files_article = [];
-      for (var i = 0; i < filesAll.length; i++) {
-        if (filesAll[i].f_key == this.thisAcademic_Article.aa_file_key) {
-          files_article.push(filesAll[i]);
-        }
+      var academic_article_show = this.academic_article_data
+      if(this.check_delete_file == false){
+        this.academic_articleE = this.academic_article_data
       }
-      if (files_article.length != 0) {
-        return files_article;
-      } else {
-        return false;
-      }
+      return academic_article_show;
     },
     academic_article_category() {
-      return this.$store.getters.getAcademic_Article_Category;
+      if(this.data_load_category==false){
+        axios.get(this.$store.getters.getBase_Url+"Academic_article/get_all_academic_article_category")
+        .then(response => {
+            // console.log(response)
+            this.academic_article_category_all = response.data
+        })
+        this.data_load_category=true
+      }
+        var category_all = this.academic_article_category_all
+      return category_all
     },
     the_user() {
       var user = this.$store.getters.getThe_User;
