+++ client/src/view/Admin/Articles/article_edit.vue
@@ -1,20 +1,20 @@
 <template>
   <div class="container">
     <center>
-      <h4 class="header">แก้ไขบทความ</h4>
+      <h4 class="header">แก้ไข บทความ</h4>
     </center>
     <div class="row mt-5" v-if="thisArticle && the_user">
       <div class="col-lg-2 col-xs-12"></div>
       <div class="col-lg-8 col-xs-12">
         <center>
-          <img v-if="url" :src="url" max-width="250px" height="250px" />
-          <img v-else :src="getImgUrl(thisArticle.a_image)" max-width="250px" height="250px" />
+          <img v-if="url" :src="url" width="100%" />
+          <img v-else :src="getImgUrl(thisArticle.a_image)" width="100%" />
         </center>
         <br />
         <br />
         <button
           type="button"
-          class="form-control btn-success col-lg-12"
+          class="form-control btn-success btn col-lg-6 block-center"
           @click="ChooseFilesImage"
         >เปลี่ยนรูป</button>
         <br />
@@ -42,7 +42,16 @@
               :value="ac.ac_id"
             >{{ ac.ac_title }}</option>
           </select>
-          <br />รายละเอียด
+          <br>
+          วิดีโอที่เกี่ยวข้อง
+          <input
+            type="text"
+            v-model="articleE.a_video_link"
+            class="form-control"
+            placeholder="วิดีโอ ที่เกี่ยวข้อง #หากไม่มีไม่ต้องใส่"
+          />
+          <br />
+          รายละเอียด
           <textarea
             v-model="articleE.a_detail"
             class="form-control textarea"
@@ -50,8 +59,8 @@
             rows="6"
           ></textarea>
           <br />
-          <div v-if="thisFiles != null" v-for="(file,run) in thisFiles">
-            <button type="button" class="btn btn-danger" @click="RemoveFile(file.f_id)">delete</button>
+          <div v-if="article_data_file != null" v-for="(file,run) in article_data_file">
+            <button type="button" class="btn btn-danger" @click="RemoveFile(file.f_id,file.f_title)">delete</button>
             <a :href="loadFile(file.f_name)" download>Dowload File</a>
             {{file.f_title}}
             <br />
@@ -111,6 +120,7 @@
   </div>
 </template>
 <script>
+import axios from "axios";
 export default {
   data() {
     return {
@@ -119,7 +129,15 @@ export default {
       url: null,
       fileimage: "",
       files: [],
-      max_size_file: 0
+      max_size_file: 0,
+
+      article_category_all:'',
+      data_load_category:false,
+      
+      article_data:'',
+      article_data_file:'',
+      data_load:false,
+      check_delete_file:false
     };
   },
   methods: {
@@ -171,11 +189,31 @@ export default {
       this.file_title.splice(index, 1);
     },
     // delete file
-    RemoveFile: function(fileID) {
+    RemoveFile: function(fileID,fileName) {
       var FD_delete = new FormData();
       FD_delete.append("fileID", fileID);
       FD_delete.append("creator", JSON.stringify(this.$store.state.log_on));
-      this.$store.dispatch("Delete_File", FD_delete);
+      
+      this.$swal({
+          title: "Are you sure?",
+          text: "You want delete this Filse Name "+fileName,
+          icon: "warning",
+          buttons: true,
+          dangerMode: true,
+        })
+        .then((willDelete) => {
+        if (willDelete) {
+          this.$store.dispatch("Delete_File", FD_delete);
+          setTimeout(() => {
+            this.check_delete_file=true
+            this.data_load=false
+          },100);
+          swal({title: "Delete Success.",icon: "success",});
+        } else {
+          //
+        }
+      })
+      
     },
     // submit
     submitArticle() {
@@ -205,32 +243,33 @@ export default {
       return this.$store.getters.getPath_Files;
     },
     thisArticle() {
-      var articleAll = this.$store.getters.getArticle;
-      var article;
-      for (var i = 0; i < articleAll.length; i++) {
-        if (articleAll[i].a_id == this.$route.params.ArticleID) {
-          article = articleAll[i];
-        }
+      var articlesID = this.$route.params.ArticleID
+      if(this.data_load==false){
+        axios.get(this.$store.getters.getBase_Url+'Article/get_this_article/'+articlesID)
+        .then(response => {
+            // console.log(response.data)
+            this.article_data = response.data[0][0],
+            this.article_data_file = response.data[1]
+        })
+        this.data_load = true
       }
-      this.articleE = article;
-      return article;
-    },
-    thisFiles() {
-      var filesAll = this.$store.getters.getFiles;
-      var files_article = [];
-      for (var i = 0; i < filesAll.length; i++) {
-        if (filesAll[i].f_key == this.thisArticle.a_file_key) {
-          files_article.push(filesAll[i]);
-        }
+      var article_show = this.article_data
+      if(this.check_delete_file == false){
+        this.articleE = this.article_data
       }
-      if (files_article.length != 0) {
-        return files_article;
-      } else {
-        return false;
-      }
+      return article_show;
     },
     article_category() {
-      return this.$store.getters.getArticle_Category;
+      if(this.data_load_category==false){
+        axios.get(this.$store.getters.getBase_Url+"Article/get_all_article_category")
+        .then(response => {
+            // console.log(response)
+            this.article_category_all = response.data
+        })
+        this.data_load_category=true
+      }
+        var category_all = this.article_category_all
+      return category_all
     },
     the_user() {
       var user = this.$store.getters.getThe_User;
