+++ client/src/view/Admin/News_Activities/news_activity_edit.vue
@@ -6,13 +6,13 @@
     <div class="row mt-5" v-if="thisNews && the_user">
       <div class="col-lg-2 col-xs-12"></div>
       <div class="col-lg-8 col-xs-12">
-        <img v-if="url" :src="url" class="admin-img" />
-        <img v-else :src="getImgUrl(thisNews.n_image)" class="admin-img" />
+        <img v-if="url" :src="url" class=" " width="100%"/>
+        <img v-else :src="getImgUrl(thisNews.n_image)" width="100%" class=" " />
         <br />
         <br />
         <button
           type="button"
-          class="form-control btn-success col-lg-12"
+          class="form-control btn-success col-lg-8 block-center"
           @click="ChooseFilesImage"
         >เปลี่ยนรูปภาพ</button>
         <br />
@@ -32,7 +32,16 @@
             placeholder="หัวเรื่อง"
             required
           />
-          <br />รายละเอียด
+          <br />
+          วิดีโอที่เกี่ยวข้อง
+          <input
+            type="text"
+            v-model="newsE.n_video_link"
+            class="form-control"
+            placeholder="วิดีโอที่เกี่ยวข้อง #หากไม่มีไม่ต้องใส่"
+          />
+          <br />
+          รายละเอียด
           <textarea
             v-model="newsE.n_detail"
             class="form-control textarea"
@@ -40,8 +49,8 @@
             rows="6"
           ></textarea>
           <br />
-          <div v-if="thisFiles != null" v-for="(file,run) in thisFiles">
-            <button type="button" class="btn btn-danger" @click="RemoveFile(file.f_id)">delete</button>
+          <div v-if="news_data_file != null" v-for="(file,run) in news_data_file">
+            <button type="button" class="btn btn-danger" @click="RemoveFile(file.f_id,file.f_title)">delete</button>
             <a :href="loadFile(file.f_name)" download>Dowload File {{run+1}}</a>
             {{file.f_title}}
             <br />
@@ -100,6 +109,7 @@
   </div>
 </template>
 <script>
+import axios from "axios";
 export default {
   data() {
     return {
@@ -108,7 +118,13 @@ export default {
       url: null,
       fileimage: "",
       files: [],
-      max_size_file: 0
+      max_size_file: 0,
+
+      data_load:false,
+      text_alert:'',
+      news_data:'',
+      news_data_file:'',
+      check_delete_file:false
     };
   },
   methods: {
@@ -160,11 +176,32 @@ export default {
       this.file_title.splice(index, 1);
     },
     //delete file
-    RemoveFile: function(fileID) {
+    RemoveFile: function(fileID,fileName) {
       var FD_delete = new FormData();
       FD_delete.append("fileID", fileID);
       FD_delete.append("creator", JSON.stringify(this.$store.state.log_on));
-      this.$store.dispatch("Delete_File", FD_delete);
+
+      this.$swal({
+          title: "Are you sure?",
+          text: "You want delete this Filse Name "+fileName,
+          icon: "warning",
+          buttons: true,
+          dangerMode: true,
+        })
+        .then((willDelete) => {
+        if (willDelete) {
+          this.$store.dispatch("Delete_File", FD_delete);
+          setTimeout(() => {
+            this.check_delete_file=true
+            this.data_load=false
+          },100);
+          swal({title: "Delete Success.",icon: "success",});
+        } else {
+          //
+        }
+      })
+      
+      
     },
     //submit
     submitNews() {
@@ -194,34 +231,31 @@ export default {
       return this.$store.getters.getPath_Files;
     },
     thisNews() {
-      var newsAll = this.$store.getters.getNews;
-      var news;
-      for (var i = 0; i < newsAll.length; i++) {
-        if (newsAll[i].n_id == this.$route.params.NewsID) {
-          news = newsAll[i];
+      var newsID = this.$route.params.NewsID;
+      if(this.data_load==false){
+        axios.get(this.$store.getters.getBase_Url+'News/get_this_news/'+newsID)
+        .then(response => {
+            // console.log(response.data),
+            this.news_data = response.data[0][0],
+            this.news_data_file = response.data[1]
+        })
+        this.data_load = true
+        if(this.news_data.length == 0){
+          setTimeout(() => {
+            this.text_alert = 'This Page No Data.'
+          },1000);
         }
       }
-      this.newsE = news;
-      return news;
-    },
-    thisFiles() {
-      var filesAll = this.$store.getters.getFiles;
-      var files_news = [];
-      for (var i = 0; i < filesAll.length; i++) {
-        if (filesAll[i].f_key == this.thisNews.n_file_key) {
-          files_news.push(filesAll[i]);
-        }
+      var news_show = this.news_data
+      if(this.check_delete_file == false){
+        this.newsE = this.news_data
       }
-      if (files_news.length != 0) {
-        return files_news;
-      } else {
-        return null;
-      }
+      return news_show;
     },
     the_user() {
       var user = this.$store.getters.getThe_User;
       if (user.m_status != "admin") {
-        this.$router.go(-1);
+        // this.$router.go(-1);
       }
       return user;
     }
